[#22] Move files for directory structures
