+++ server/common/inversify/createContainer.ts
@@ -1,9 +1,9 @@
 import { Container } from 'inversify';
 import { NextApplication } from '../nextjs/NextApplication';
 
-import "../../common/api/CommonController";
 import "../../mother/api/MotherController";
 import "../../mother/notice/api/NoticeController";
+import "../api/CommonController";
 
 export const createContainer = () => {
   const container = new Container();
