+++ server/common/nextjs/NextApplication.ts
