--- server/main/index.ts
@@ -1,16 +1,0 @@
-import "reflect-metadata";
-
-import { createExpressApp } from "./common/inversify/createExpressApp";
-import { createInversifyContainer } from "./common/inversify/createInversifyContainer";
-import { NextApplication } from "./common/nextjs/NextApplication";
-import { logger } from "./common/utils";
-import { errorHandlers } from "./errorHandlers";
-
-const PORT = 3000;
-
-const nextApplication = new NextApplication();
-
-nextApplication.run().then(() => {
-  createExpressApp(createInversifyContainer(), errorHandlers).listen(PORT);
-  logger.log("info", `server is running on port:${PORT}`);
-});
